RAG/requirements.txt
snowflake-connector-python
python-dotenv
pandas
trulens-core 
trulens-providers-cortex 
trulens-connectors-snowflake 
snowflake-sqlalchemy
trulens 